add: 유저가 사장님인지 확인하는 로직 & 테스트 추가

+++ src/controllers/store.controller.js
@@ -7,11 +7,17 @@ export class StoreController {
     try {
       // 로그인 인증 인가 미들웨어를 통해서 로그인한 유저의 정보를 가져와야함 그래서 크리에이트할떄 유저 정보를 입력하는 식으로 유저 ID, 유저 주소 이럴 느낌
       //const userId = req.cookie;
+      //const type = req.user;
 
-      const { userId, storeName, foodType } = req.body; // insomnia 테스트를 위해 userId를 바디값으로 받음
+      const { userId, storeName, foodType, type } = req.body; // insomnia 테스트를 위해 userId를 바디값으로 받음
 
       if (!storeName || !foodType) {
         throw new Error('InvalidParamsError');
+      }
+
+      // 사장과 손님 확인 작업
+      if (type !== true) {
+        throw new Error('해당하는 유저는 사장님이 아닙니다.');
       }
 
       const createdStore = await this.storeService.createStore(userId, storeName, foodType);
@@ -25,7 +31,12 @@ export class StoreController {
   updateStore = async (req, res, next) => {
     try {
       //   const userId = req.cookie;
-      const { userId, storeName, foodType } = req.body; // insomnia 테스트를 위해 userId를 바디값으로 받음
+      const { userId, storeName, foodType, type } = req.body; // insomnia 테스트를 위해 userId를 바디값으로 받음
+
+      // 사장과 손님 확인 작업
+      if (type !== true) {
+        throw new Error('해당하는 유저는 사장님이 아닙니다.');
+      }
 
       const updatedStore = await this.storeService.updateStore(userId, storeName, foodType);
       return res.status(200).json({ data: updatedStore });
@@ -38,7 +49,13 @@ export class StoreController {
     try {
       // 인증 인가 부분이 없어서 임시로 body로 추가
       // const { userId } = req.user;
-      const { storeName, userId } = req.body;
+      // const type = req.user;
+      const { storeName, userId, type } = req.body;
+
+      // 사장과 손님 확인 작업
+      if (type !== true) {
+        throw new Error('해당하는 유저는 사장님이 아닙니다.');
+      }
 
       // 서비스 계층에 구현된 deleteStore 로직을 실행합니다.
       const deletedStore = await this.storeService.deleteStore(userId, storeName);
